+++ server/src/main/java/com/krillsson/sysapi/core/domain/network/Connectivity.kt
@@ -3,5 +3,6 @@ package com.krillsson.sysapi.core.domain.network
 data class Connectivity(
     val externalIp: String?,
     val previousExternalIp: String?,
+    val localIp: String?,
     val connected: Boolean
 )
