+++ server/src/main/java/com/krillsson/sysapi/core/history/db/Connectivity.kt
@@ -12,6 +12,7 @@ class Connectivity(
     val id: UUID,
     val externalIp: String?,
     val previousExternalIp: String?,
+    val localIp: String?,
     val connected: Boolean
 )
