+++ server/src/main/resources/db/migration/V2__Migrate_connectivty_table.sql
@@ -1,0 +1,1 @@
+ALTER TABLE Connectivity ADD COLUMN localIp VARCHAR(255) NULL;
